made error message more helpful and shortened docstring
The original docstring contained two entire base64 versions of the LICENSE file, which made it extremely long and let the two lines spike out. This proposal shortens these by a good bit, so they don't occupy as much space and are better readable.

As well, the error message in case of an incompatible type now returns the faulty type of the argument that was given to make the possible origin of any errors in the future more clear.