Add OLLAMA_HOST blurb to README
Without this it's opaque that the library is making HTTP requests,
or to where it might be making them.

The only place I could find how to change the localhost:11434 behavior
was in the Ollama FAQs. If there is a way to set it besides OLLAMA_HOST I would love to know!